v0_8_2

+++ src/main/java/be/access/UserDAO.java
@@ -2,10 +2,12 @@ package be.access;
 
 import be.entity.User;
 
+import java.math.BigDecimal;
+
 public interface UserDAO {
 
     User getUserByLogin(String login);
     User getUserById(Long id);
     void registrationUser(User newUser);
-    void setBalance(Long id, Double balance);
+    void setBalance(Long id, BigDecimal balance);
 }
