+++ src/main/java/be/business/UserBusinessService.java
@@ -3,10 +3,12 @@ package be.business;
 import be.entity.User;
 import be.utils.ServiceException;
 
+import java.math.BigDecimal;
+
 public interface UserBusinessService{
 
     User login(String user, String password);
     void registration(String user, String password, String address);
     User getUserById(Long id);
-    void pay(Long id, Double priceOrder) throws ServiceException;
+    void pay(Long id, BigDecimal priceOrder) throws ServiceException;
 }
